ReadingComp/requirements.txt
langgraph
python-docx
sqlalchemy
streamlit
python-dotenv
pydantic
Pillow
google-genai